chore: move tall-stack files into a dedicated directory
