+++ resources/js/tall-stack/alpine/modalComponent/index.ts
