+++ resources/js/tall-stack/alpine/modalComponent/modalComponent.ts
