+++ resources/js/tall-stack/alpine/navbarSearchComponent/index.ts
