+++ resources/js/tall-stack/alpine/navbarSearchComponent/navbarSearchComponent.ts
