+++ resources/js/tall-stack/alpine/newsCarouselComponent/index.ts
