+++ resources/js/tall-stack/alpine/newsCarouselComponent/newsCarouselComponent.ts
