+++ resources/js/tall-stack/alpine/toggleAchievementRowsComponent/index.ts
