+++ resources/js/tall-stack/alpine/tooltipComponent/middleware/cursorFlipMiddleware.ts
