+++ resources/js/tall-stack/alpine/tooltipComponent/middleware/cursorShiftMiddleware.ts
