+++ resources/js/tall-stack/alpine/tooltipComponent/state/tooltipStore.ts
