+++ resources/js/tall-stack/alpine/tooltipComponent/utils/getIsMobileIos.test.ts
