+++ resources/js/tall-stack/alpine/tooltipComponent/utils/hideTooltip.ts
