+++ resources/js/tall-stack/alpine/tooltipComponent/utils/loadDynamicTooltip.ts
