+++ resources/js/tall-stack/alpine/tooltipComponent/utils/pinTooltipToCursorPosition.test.ts
