+++ resources/js/tall-stack/alpine/tooltipComponent/utils/pinTooltipToCursorPosition.ts
