+++ resources/js/tall-stack/alpine/tooltipComponent/utils/renderTooltip.ts
