+++ resources/js/tall-stack/alpine/tooltipComponent/utils/trackTooltipMouseMovement.ts
