+++ resources/js/tall-stack/alpine/tooltipComponent/utils/updateTooltipPosition.ts
