+++ resources/js/tall-stack/app.ts
@@ -5,16 +5,6 @@
 
 // eslint-disable-next-line @typescript-eslint/ban-ts-comment -- This actually works in a TALL stack app with Livewire 3.
 // @ts-ignore
-import { Alpine, Livewire } from '../../vendor/livewire/livewire/dist/livewire.esm';
-import {
-  linkifyDirective,
-  modalComponent,
-  navbarSearchComponent,
-  newsCarouselComponent,
-  toggleAchievementRowsComponent,
-  tooltipComponent,
-} from './alpine';
-import { lazyLoadModuleOnIdFound } from './lazyLoadModuleOnIdFound';
 import {
   autoExpandTextInput,
   copyToClipboard,
@@ -30,7 +20,18 @@ import {
   themeChange,
   toggleUserCompletedSetsVisibility,
   updateUrlParameter,
-} from './utils';
+} from '@/utils';
+
+import { Alpine, Livewire } from '../../../vendor/livewire/livewire/dist/livewire.esm';
+import {
+  linkifyDirective,
+  modalComponent,
+  navbarSearchComponent,
+  newsCarouselComponent,
+  toggleAchievementRowsComponent,
+  tooltipComponent,
+} from './alpine';
+import { lazyLoadModuleOnIdFound } from './lazyLoadModuleOnIdFound';
 
 // livewire_hot_reload();
