+++ resources/js/tall-stack/dynamic/reorderSiteAwards/cancelAutoscroll.ts
